app/backend/requirements.txt
azure-identity==1.13.0
Flask==2.2.5
langchain==0.0.187
openai==0.27.8
tiktoken==0.3.0
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
Flask-SQLAlchemy
flask-bcrypt
python-dotenv
flask-session
redis
flask-cors